python_service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.6.0
supabase>=2.3.0
google-generativeai>=0.3.2
litellm>=1.17.0
langgraph>=0.0.10
langchain>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.3
pdfplumber>=0.10.3
sentence-transformers>=2.3.1
# OCR ve Resim İşleme İçin (Versiyonu serbest bıraktık):
pytesseract
Pillow>=10.2.0